refactor: drop support for docrtine 3.6
