+++ composer.json
@@ -25,7 +25,7 @@
         "brick/math": "^0.12 || ^0.13 || ^0.14 || ^0.15 || ^0.16 || ^0.17",
         "coduo/php-humanizer": "^5.0",
         "cuyz/valinor": "^2.4",
-        "doctrine/dbal": "^3.6 || ^4.0",
+        "doctrine/dbal": "^4.0",
         "elasticsearch/elasticsearch": "^7.6|^8.0",
         "google/apiclient": "^2.13",
         "google/protobuf": "^4.0 || ^5.0",
