+++ src/lib/doctrine-dbal-bulk/composer.json
@@ -12,7 +12,7 @@
     ],
     "require": {
         "php": "~8.3.0 || ~8.4.0 || ~8.5.0",
-        "doctrine/dbal": "^3.6 || ^4.0",
+        "doctrine/dbal": "^4.0",
         "flow-php/types": "self.version"
     },
     "autoload": {
